requirements.txt
flask
gunicorn
google-cloud-storage
opencolorio